download the json file synchronously
fixy fix fix

+++ ew.js
@@ -1,18 +1,12 @@
 const isRoot = process.getuid && process.getuid() === 0;
 const fs = require('fs-extra');
 const download = require('download-file');
+const download2 = require(`download-file-sync`);
 const targz = require('targz');
-
 if (!isRoot) return console.log("You must be root to excecute this command!");
 if (!fs.existsSync("/etc/ew")) fs.mkdirSync("/etc/ew");
 if (!fs.existsSync("/etc/ew/installed")) fs.mkdirSync("/etc/ew/installed");
 //make an archive to help with other functions
-if (!fs.existsSync("/etc/ew/archive")) fs.mkdirSync("/etc/ew/archive");
-
-if (!fs.existsSync("/etc/ew/source")) {
-    fs.mkdirSync("/etc/ew/source");
-    fs.writeFileSync("/etc/ew/sources.list");
-}
 //if (fs.existsSync("/etc/ew/lock")) return console.log("/etc/ew/lock exists, cancelling.");
 //fs.writeFileSync('/etc/ew/lock', 'ew package manager lock\n');
 //console.log("Created lock file.");
@@ -24,6 +18,7 @@ if (process.argv[2] === '-r') updatePackageList();
 if (process.argv[2] === '-i') install(process.argv[3]);
 if (process.argv[2] === '-u') uninstall(process.argv[3]);
 if (process.argv[2] === '-h') help();
+if (process.argv[2] === '-re') removeew();
 
 function updatePackageList() {
     process.stdout.write("Updating package list... ");
@@ -34,41 +29,34 @@ function updatePackageList() {
 };
 
 function help() {
-    console.log("here are the list of commands available for ew package manager:");
-    console.log(" ");
-    console.log("-r: reload package list");
-    console.log("-i: installs a package");
-    console.log("-u: uninstalls a package");
-    console.log("-h: prints list of commands");
+    console.log("here are the list of commands available for ew package manager:\n \n -r: reload package list\n -i: installs a package\n -u: uninstalls a package\n -h: prints list of commands\n");
 }
 
 function install(package) {
     var packageList = require("/etc/ew/packages.json");
-    var packageURL = packageList[package]
     var packageURL = packageList[package];
-    console.log(packageURL);
     process.stdout.write(`Downloading ${package}... `);
     download(packageURL, { directory: "/tmp", filename: `${package}.ew.tar.gz` }, function() {
         targz.decompress({ src: `/tmp/${package}.ew.tar.gz`, dest: '/' }, function() {
-            fs.unlinkSync(`/tmp/${package}.ew.tar.gz`);
-            var jsondata = fs.readFileSync(`/etc/ew/installed/${package}.json`);
-            var json = JSON.parse(jsondata);
-            if (json["depends"]) {
-                let dependscount = json.depends.length;
-                if (dependscount === 1) console.log("this package requires " + dependscount + " dependency, installing now...");
-                else if (dependscount === 0) return console.log("Warning: json file has a depends field, but its empty.");
-                else console.log("this package requires" + dependscount + "dependencies");
-                for (let dependencies = json.depends; dependscount > 0; dependscount--) {
-                    dependencies = JSON.stringify(dependencies);
-                    dependencies = dependencies.replace(/[\[\]"]/g, "")
-                    install(dependencies);
-                }
-            }
+            if (fs.existsSync(`/tmp/${package}.ew.tar.gz`)) fs.unlinkSync(`/tmp/${package}.ew.tar.gz`);
         });
-        process.stdout.write("Done!\n");
-        process.stdout.write(`Installed ${package}!\n`);
     });
-    download(`https://ew.cumbox.best/packageInfo/${package}.json`, { directory: "/etc/ew/installed", filename: `${package}.json` });
+    let jsond = download2(`https://ew.cumbox.best/packageInfo/${package}.json`);
+    fs.writeFileSync(`/etc/ew/installed/${package}.json`, jsond);
+    process.stdout.write("done!\n");
+    process.stdout.write(`Successfully installed ${package}!\n`);
+    const json = require(`/etc/ew/installed/${package}.json`);
+    if (json["depends"]) {
+        let dependscount = json.depends.length;
+        if (dependscount === 1) console.log("this package requires " + dependscount + " dependency, installing now...");
+        else if (dependscount === 0) return console.log("Warning: json file has a depends field, but its empty.");
+        else console.log("this package requires" + dependscount + "dependencies");
+        for (let dependencies = json.depends; dependscount > 0; dependscount--) {
+            dependencies = JSON.stringify(dependencies);
+            dependencies = dependencies.replace(/[\[\]"]/g, "")
+            install(dependencies);
+        }
+    }
 };
 
 function uninstall(package) {
@@ -85,5 +73,16 @@ function uninstall(package) {
     fs.unlinkSync(`/etc/ew/installed/${package}.json`);
     process.stdout.write(`done!\n`);
     process.stdout.write(`Successfully  uninstalled ${package}!\n`);
+}
+
+function removeew() {
+    process.stdout.write("removing ewpm...\n");
+    if (fs.existsSync(`/etc/ew/installed/*.json`)) return console.log("Warning: you have packages installed, please remove them");
+    if (fs.existsSync(`/etc/ew/`)) console.log("ew dir has been found");
+    if (fs.existsSync(`/etc/ew/packages.json`)) fs.removeSync("/etc/ew/packages.json");
+    if (fs.existsSync("/etc/ew/archive")) fs.removeSync("/etc/ew/archive");
+    if (fs.existsSync("/etc/ew/source/sources.list")) fs.removeSync("/etc/ew/source/sources.list");
+    if (fs.existsSync("/etc/ew/source")) fs.removeSync("/etc/ew/source");
+    return console.log("ewpm has been successfully removed!");
 }
 //fs.unlinkSync('/etc/ew/lock');
